requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   openai
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
defusedxml==0.7.1
    # via youtube-transcript-api
distro==1.9.0
    # via openai
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   langsmith
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.8.2
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain-core==0.3.23
    # via
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.2.11
    # via transcripts (pyproject.toml)
langchain-text-splitters==0.3.2
    # via transcripts (pyproject.toml)
langsmith==0.1.147
    # via langchain-core
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   tweepy
openai==1.57.2
    # via langchain-openai
orjson==3.10.12
    # via langsmith
packaging==24.2
    # via langchain-core
pydantic==2.10.3
    # via
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.27.1
    # via pydantic
python-dotenv==1.0.1
    # via transcripts (pyproject.toml)
pyyaml==6.0.2
    # via langchain-core
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   langsmith
    #   requests-oauthlib
    #   requests-toolbelt
    #   tiktoken
    #   tweepy
    #   youtube-transcript-api
requests-oauthlib==1.3.1
    # via tweepy
requests-toolbelt==1.0.0
    # via langsmith
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tenacity==9.0.0
    # via langchain-core
tiktoken==0.8.0
    # via langchain-openai
tqdm==4.67.1
    # via openai
tweepy==4.14.0
    # via transcripts (pyproject.toml)
typing-extensions==4.12.2
    # via
    #   anyio
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.3
    # via requests
youtube-transcript-api==0.6.3
    # via transcripts (pyproject.toml)
